+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `tooncher.LoginSuccessful`
   - `tooncher.TOONTOWN_ENGINE_DEFAULT_PATH`
   - `tooncher.api_request`
+  - `tooncher.login`
 - `start_engine` & `launch`: expected `isinstance(engine_path, pathlib.Path)`
   (instead of `str`)